requirements.txt
attrs==19.1.0
aws-cdk.assets==1.5.0
aws-cdk.aws-apigateway==1.5.0
aws-cdk.aws-applicationautoscaling==1.5.0
aws-cdk.aws-autoscaling==1.5.0
aws-cdk.aws-autoscaling-common==1.5.0
aws-cdk.aws-autoscaling-hooktargets==1.5.0
aws-cdk.aws-certificatemanager==1.5.0
aws-cdk.aws-cloudformation==1.5.0
aws-cdk.aws-cloudfront==1.5.0
aws-cdk.aws-cloudwatch==1.5.0
aws-cdk.aws-codebuild==1.5.0
aws-cdk.aws-codecommit==1.5.0
aws-cdk.aws-codedeploy==1.5.0
aws-cdk.aws-codepipeline==1.5.0
aws-cdk.aws-codepipeline-actions==1.5.0
aws-cdk.aws-ec2==1.5.0
aws-cdk.aws-ecr==1.5.0
aws-cdk.aws-ecr-assets==1.5.0
aws-cdk.aws-ecs==1.5.0
aws-cdk.aws-elasticloadbalancing==1.5.0
aws-cdk.aws-elasticloadbalancingv2==1.5.0
aws-cdk.aws-events==1.5.0
aws-cdk.aws-events-targets==1.5.0
aws-cdk.aws-iam==1.5.0
aws-cdk.aws-kms==1.5.0
aws-cdk.aws-lambda==1.5.0
aws-cdk.aws-logs==1.5.0
aws-cdk.aws-route53==1.5.0
aws-cdk.aws-route53-targets==1.5.0
aws-cdk.aws-s3==1.5.0
aws-cdk.aws-s3-assets==1.5.0
aws-cdk.aws-secretsmanager==1.5.0
aws-cdk.aws-servicediscovery==1.5.0
aws-cdk.aws-sns==1.5.0
aws-cdk.aws-sns-subscriptions==1.5.0
aws-cdk.aws-sqs==1.5.0
aws-cdk.aws-ssm==1.5.0
aws-cdk.aws-stepfunctions==1.5.0
aws-cdk.core==1.5.0
aws-cdk.custom-resources==1.5.0
aws-cdk.cx-api==1.5.0
aws-cdk.region-info==1.5.0
cattrs==0.9.0
jsii==0.15.1
publication==0.0.3
python-dateutil==2.8.0
# Editable Git install with no remote (serverless-pipeline==0.0.1)
-e /Users/scottschmidt/Projects/playground/serverless-pipeline/serverless_pipeline
six==1.12.0
typing-extensions==3.7.4
